[REFACTO] Add tests + extremly large number handling

+++ src/main/java/com/unarbre/stringcalculator/StringCalculator.java
@@ -3,12 +3,14 @@ package com.unarbre.stringcalculator;
 import com.unarbre.stringcalculator.parser.IntParser;
 import com.unarbre.stringcalculator.parser.separators.SeparatorStrategyFactory;
 
+import java.math.BigInteger;
+
 final class StringCalculator {
 
     private final SeparatorStrategyFactory separatorStrategyFactory = new SeparatorStrategyFactory();
 
 
-    public int add(String rawInput) {
+    public BigInteger add(String rawInput) {
         var userInput = new UserInput(rawInput);
         var separatorStrategy = this.separatorStrategyFactory.get(userInput);
         var inputParser = new IntParser(separatorStrategy);
